requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
Django==4.2.7
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
greenlet==3.0.1
h11==0.14.0
msgpack==1.0.7
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg==3.1.12
pynvim==0.4.3
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
uvicorn==0.24.0.post1
